Add license and copyright info
Enables and uses the macros plugin so license info (namely the full license
name) so we can write out that info once and mention it anywhere using macros.
Had to do a bunch of fixes for the Wiki syntax to not confuse that.

Also enables and uses the abbr and pymdownx.snippets extensions so we can use
abbreviations like "HIG" and have a global registry of the full terms. The full
terms appear on mouse hover then.